+++ packages/client/src/links/httpLink.ts
@@ -1,5 +1,5 @@
 import { AnyRouter } from '@trpc/server';
-import { TRPCClientError } from '../createTRPCClient';
+import { TRPCClientError } from '../internals/TRPCClientError';
 import { transformRPCResponse } from '../internals/transformRPCResponse';
 import { httpRequest } from '../internals/httpRequest';
 import { HttpLinkOptions, TRPCLink } from './core';
