+++ packages/client/src/links/wsLink.ts
@@ -4,7 +4,7 @@ import {
   TRPCRequest,
   TRPCResult,
 } from '@trpc/server/rpc';
-import { TRPCClientError } from '../createTRPCClient';
+import { TRPCClientError } from '../internals/TRPCClientError';
 import { ObservableCallbacks, UnsubscribeFn } from '../internals/observable';
 import { retryDelay } from '../internals/retryDelay';
 import { TRPCLink } from './core';
