+++ packages/server/test/links.test.ts
@@ -3,7 +3,8 @@ import { waitFor } from '@testing-library/dom';
 import AbortController from 'abort-controller';
 import fetch from 'node-fetch';
 import { z } from 'zod';
-import { createTRPCClient, TRPCClientError } from '../../client/src';
+import { createTRPCClient } from '../../client/src';
+import { TRPCClientError } from '../../client/src/internals/TRPCClientError';
 import { executeChain } from '../../client/src/internals/executeChain';
 import { LinkRuntimeOptions, OperationLink } from '../../client/src/links/core';
 import { httpBatchLink } from '../../client/src/links/httpBatchLink';
